Inference: Transform images to binary format before saving them, convert them as float when reading

+++ robot/real/Camera.py
@@ -14,10 +14,9 @@ class Camera(Component):
     mask_obstacle = None
 
     def execute(self):
-
         # Consume masks produced by inference process
-        self.mask_line = np.load(self.mask_line_file_path)
-        self.mask_obstacle = np.load(self.mask_obstacle_file_path)
+        self.mask_line = np.load(self.mask_line_file_path).astype(float)
+        self.mask_obstacle = np.load(self.mask_obstacle_file_path).astype(float)
 
 
 if __name__ == '__main__':
